project code
CPU schedules N processes which arrive at different time intervals and each process allocated the CPU for a specific user input time unit, processes are scheduled using  a  preemptive round  robin  scheduling  algorithm.  Each process must is be assigned a numerical priority, with a higher number indicating a higher relative  priority.  In addition priority 0.  The  length  of  a time quantum time quantum for processing.
If is T units,  where  T a  process is is to the  processes  one  task  has the   custom time considered as preempted by  a  higher preempted  process is placed  at  the  end  of  t priority  process,  the he  queue.  Design  a  scheduler  so that the task with priority 0 does not  starve  for  resources  and  gets  the CPU at some compute waiting time, turn around.